backend/requirements.txt
elasticsearch==9.1.2
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4
